+++ Source/Mosa.Test.Collection/ConvI4Tests.cs
@@ -6,17 +6,17 @@ namespace Mosa.Test.Collection
 	{
 		public static bool ConvI4_I1(int expect, sbyte a)
 		{
-			return expect == ((int)a);
+			return expect == a;
 		}
 
 		public static bool ConvI4_I2(int expect, short a)
 		{
-			return expect == ((int)a);
+			return expect == a;
 		}
 
 		public static bool ConvI4_I4(int expect, int a)
 		{
-			return expect == ((int)a);
+			return expect == a;
 		}
 
 		public static bool ConvI4_I8(int expect, long a)
